Fix all warnings for Linux release build
"Unused but set" is still suppressed, Windows is broken by that flag
since Ubuntu's packaged mingw32 g++ is quite a bit older, and there
are probably a bunch of warnings hiding behind DEMO-define code.

Mostly making sure things are initialized for theoretical cases where
none of the if/elseif branches are taken. A few syntax/type cleanups
for disambiguation. A bunch of checks of fread returns, most of which
throw, but one indicates that one of the sprites in the release game
data is actually truncated as far as the game is expecting. Now at
least zero-initializes the leftover part explicitly.