Clean up set-but-unused variables
Some of these are used by fragments of commented-out code, so tried to
preserve enough state for if anyone wants to experiment with bringing
that code back. The rest just looked like editing detritus.

Fixes the cross-compile with the older G++ again.

Also fix a misreported error about short reads when loading settings;
missed that it was a case where the size/count arguments were reversed,
but they make sense in this order.