Fix build for cross-compiled Windows as well
Add a quick script that fires off both builds for testing them both.

Fix warnings that were upsetting the older G++ in the cross-compiler.
Some mising end-of-file newlines, one embedded null in a C string that
was hiding behind a WINDOWS define, and some cases where it wasn't able
to determine that a variable would be initialized before use that the
newer, native G++ had managed to figure out were safe. One type
widening to match the declared return value and resolve an always-
satisfied inequality (char < 128).

Now that we're warning-clean (at least for release, and DEMO probably
isn't going to get much attention), make warnings fatal so we stay that
way.