Support window resize and fullscreen
Make the window resizable, and reassign a scaler when this happens. On
the fly resizing now works.

Handle F11 to toggle the apparently previously unused gfx_fullscreen
toggle, and respect it along with window resizing.

SDL 1.2 doesn't have SetWindowMinimumSize, so you can currently try to
shrink the window and thwart the game's attempts to reassert a sensible
size, but it doesn't crash (...any more).

Fix the fix to the integer scale factor check. We can never downscale,
and trying to with code that's not expecting it can cause hilarious
segfaults.