Fix up and enable the HQ arbitrary scaler
Downside: it was written for 320x240 quasi-16-color graphics; SAIS'
640x480 256-colour animations are absolutely smashing the blend cache,
and performance is abysmal.

But it does give a working 8-to-32-bit scaler for fullscreen, now it's
been convinced to work with palettes as input.

Probably better to put it aside and do a variant of the regular
arbitrary scaler which is also palette-literate.